+++ examples/private_site_with_login.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require 'medusa'
+require 'http/cookie'
 require 'net/http'
 require 'uri'
-require 'webrick/cookie'
 
 module PrivateSiteWithLogin
   module_function
@@ -22,12 +22,12 @@ module PrivateSiteWithLogin
     end
 
     cookies = response.get_fields('Set-Cookie').to_a.flat_map do |header|
-      WEBrick::Cookie.parse_set_cookies(header)
+      ::HTTP::Cookie.parse(header, uri)
     end
 
     raise 'Login response did not set a session cookie' if cookies.empty?
 
-    cookies.to_h { |cookie| [cookie.name, cookie.value] }
+    cookies
   end
 
   def crawl(start_url:, login_url:, username:, password:, **options, &block)